fix: Apply feedback
